stackview: fix slow highlight behavior
fix the slowness caused by stackview highlight that happens when
going down the stack when the stack tree contains wide levels
(it doesn't have to be deep for it to happen).

the main performance bottleneck was due to the use of deepcopy during
every path evaluation in the tree traversal.
when a path included 'wide nodes' (nodes with many children), repeatedly
performing deep copies in each iteration caused significant slowdowns,
often making NVIM appear unresponsive

the fix is indentation-based and only operates on the visible lines in
the buffer and doesn't depend on the structure of the tree:
all it does is to highlight the lines above the cursor every time the
indentation gets smaller.

Signed-off-by: aviyae <aviyae42@gmail.com>