quickfix/location: fix unsaved history
new cscope commands are expected to have a new list on the stack
and not to override the current list - preventing from returning to
it later.

fix that by removing the "r" ("replace") option from the
setqflist/setloclist command and by that using the default option
instead, which creates a new list on top of the stack lists.

Signed-off-by: aviyae <aviyae42@gmail.com>