StackView: Allow scaling of stack view window size
Implemented width_scale and height_scale options for stack view window.

This replaces the previous fixed scaling factors with user-configurable
options, and provides default values if the user does not specify them.
Values are expressed as a fraction of the total screen size.

Values are always clamped to be between > 0 and <= 1 to prevent invalid
sizes. If out of range values are provided, defaults are used.

The implementation ensures that the stack view window is always
centered on the screen.

Signed-off-by: Izhar Ameer Shaikh <izharits@gmail.com>