picker: snacks: allow passing snack layout options (#90)
Instead of hardcoding the options for snacks picker, allow passing them
via opts.cscope.picker_opts.snacks.

This provides more flexibility to customize the snacks picker layout
according to the user's preferences.

README.md is updated with an advanced example showing how to configure a
layout with horizontal split and rounded borders, with 60% of total area
allocated to input and results, and 40% to preview.

Signed-off-by: Izhar Ameer Shaikh <izharits@gmail.com>