readme: add warning about fd -e c -e h (#92)
* Warning about fd -e c -e h

So, my goofy as simply copy-pasted the config and didn't bother trying to understand what's going. Gutentags stopped working and I couldn't figure what was happening until I realized the problem is with this specific line. So, I propose to add a warning for fellows like me who might stumble upon this problem.

* Update README.md